backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
bcrypt==4.2.1
python-multipart==0.0.20
aiofiles==24.1.0
